backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.32
psycopg[binary]==3.2.3
pydantic==2.9.1
fastapi
uvicorn[standard]
pandas==2.2.2
ijson==3.2.3
